requirements.txt
aioodbc==0.5.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
email-validator==2.1.0.post1
fastapi==0.109.0
httptools==0.6.1
httpx==0.26.0
passlib==1.7.4
pydantic==2.6.0
pyodbc==5.0.1
pytest==8.0.0
PyJWT==2.8.0
PyYAML==6.0.1
sqlmodel==0.0.14
SQLAlchemy==2.0.25
uvicorn==0.27.0.post1
websockets==12.0